+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * Raised minimum PHP version to 5.5
 * Added basic support for foreign EntityIds
 	* Added `isForeign`, `getRepository` and `getLocalPart` to `EntityId`
+	* The constructor of `EntityId` was made public
 	* Added static `EntityId::splitSerialization` and `EntityId::joinSerialization`
 	* `getNumericId` throws an exception for foreign EntityIds
 	* Added documentation for foreign EntityIds